requirements.txt
# Core dependencies for audio transcription
openai-whisper==20231117
ffmpeg-python==0.2.0

# Web backend (optional, for Flask-based web tool)
Flask==3.0.3

# Utilities (optional but recommended)
tqdm==4.66.4
numpy>=1.23.0
torch>=2.1.0
openai>=1.0.0
pytube
pandas
python-dotenv
requests
yt-dlp











